SnapshotsCompiler package renamed
